edit&delete protection & mongoose.isModified() : returns Boolean

+++ src/controllers/videoController.js
@@ -56,8 +56,16 @@ export const see = async (req, res) => {
 export const getEdit = async (req, res) => {
   const { id } = req.params;
   const video = await Video.findById(id);
+  const {
+    session: {
+      user: { _id },
+    },
+  } = req;
   if (!video) {
     return res.status(404).render("404", { pageTitle: `NOT FOUND VIDEO` });
+  }
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect("/");
   }
   return res.render("edit", { pageTitle: `Editing ${video.title}`, video });
 };
@@ -65,11 +73,22 @@ export const getEdit = async (req, res) => {
 export const postEdit = async (req, res) => {
   const { id } = req.params;
   const { title, description, hashtags } = req.body;
+  const {
+    session: {
+      user: { _id },
+    },
+  } = req;
 
-  const video = await Video.exists({ _id: id });
+  const video = await Video.findOne({ _id: id });
+
   if (!video) {
     return res.status(404).render("404", { pageTitle: `NOT FOUND VIDEO` });
   }
+
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect("/");
+  }
+
   await Video.findByIdAndUpdate(id, {
     title,
     description,
@@ -81,6 +100,20 @@ export const postEdit = async (req, res) => {
 
 export const deleteVideo = async (req, res) => {
   const { id } = req.params;
+  const {
+    session: {
+      user: { _id },
+    },
+  } = req;
+  const video = await Video.findById(id);
+  if (!video) {
+    return res.status(404).render("404", { pageTitle: `NOT FOUND VIDEO` });
+  }
+
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect("/");
+  }
+
   await Video.findByIdAndDelete(id);
   return res.redirect("/");
 };
